requirements.txt
numpy
Pillow
tensorflow-gpu==1.12.0
tabulate
tqdm
tflearn
sklearn
imutils